+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 1.1.3 - 2022-03-01
+## 1.2.0 - 2022-05-19
+Added fastapi functionality.
+Added templates for fastapi base image.
+Added testing from tapipy repo.
+Added testing Dockerfile.
+
+## 1.1.3 - 2022-05-19
 Packaged tapisservice and released on PyPi.
 
 ## 1.1.0 - 2022-03-01